requirements.txt
Django==1.8.2
<<<<<<< HEAD
django-postgrespool==0.3.0
django-toolbelt==0.0.1
django-bootstrap3==5.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.5
pytz==2015.4
fake-factory==0.5.1
ipython==3.1.0
psycopg2==2.6.1
Werkzeug==0.10.4
=======
django-extensions==1.5.5
djangorestframework==3.1.3
gnureadline==6.3.3
ipython==3.2.0
six==1.9.0
>>>>>>> d8bd4943cabe5be5d78554a315a43751f24277ca
